Add whitelist support to read/write interfaces
A whitelist is added that specifies a list of MSRs, and associated read and
write masks.  When a read is performed, only the bits in the read mask are
returned to the user.  Similarly, when a write is performed, only the bits
in the write mask are changed (we read the MSR first to read the original
values).  If the masks have no bits set, EINVAL is returned, and no read/write
is performed.

Since we are now checking permissions at MSR read/write time, the capability
check is removed from msr_open.  Whitelist checking is not in place yet for
the ioctl interface, so we add capability checking there as well.

If the user has the CAP_SYS_RAWIO capability, the module allows complete MSR
access via both the read/write and ioctl interfaces.