Fix partial write masking
The code segments that combine the previously read msr value with the value
to be written were not properly masking the previously read value.  As such,
for any MSR where some but not all of the bits were allowed to be written,
if the bit's previous value was a 1, the user could only write another 1, not
a 0.
If some bits in a msr were write-masked and others weren't, the rout