Fixes stack size warning. (#83)
The struct cpuset can be unusually large; putting it on the stack
generates a compile-time warning.  This commit uses
CONFIG_CPUMASK_OFFSTACK to determine if we can allocate struct
cpuset at runtime.  If that constant is not defined, we fall back
to the original warning-generating behavior.

I'm not sure this is a great solution, but ideally this struct
will not have to be allocated often and the performance impact
should only be in the initial latency before batch measurements.