Adds /dev/cpu/msr_safe_version device.
Catting (or just reading) the file prints out the version of
the loaded msr-safe kernel module.

It's a lot of code to just get a version number, but I suspect
many users will not otherwise have access to that information,
short of emailing the sysadmin who installed the module.  A
lighter-weight alternative would have been to put the information
into an ioctl() call, but that would require writing code just
to figure out the version number.