Committing a servicable, if flawed, version.
Debugging enabled in the Makefile.  That should be turning
off in production.

TSC (maybe) and thermal (almost certainly) MSRs are Intel-
specific.  Would like the flexibility to specify which additional
MSRs to read when.  module_param() and module_param_array() may
be the way forward, but not on this branch.

Polling is working very well.